backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
opencv-python
grad-cam
python-multipart
aiofiles
